+++ routes/adminbackend.php
@@ -1,3 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+// Connect the Controllers
+use App\Http\Controllers\ProfileController;
+
+Route::get('/backend/admin/getProfile', [ProfileController::class, 'backend_getAdminProfile']);